+++ DeckIQ.Core/Requests/FlashCards/CreateFlashCardRequest.cs
@@ -28,7 +28,8 @@ public class CreateFlashCardRequest : Request
     
     public string? CardImage { get; set; } 
     
-    public long CategoryId { get; set; }
+    public int CategoryId { get; set; }
+    
     
     public new string UserId { get; set; } = string.Empty;
 }
